moves unit tests in  __tests__ to src closes #83
