+++ minimal_redux_poc/src/__tests__/components/Display.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import Display from '../../../src/components/Display';
-import ManifestMetadata from '../../../src/components/ManifestMetadata';
-import fixture from '../../fixtures/2.json';
+import Display from '../../components/Display';
+import ManifestMetadata from '../../components/ManifestMetadata';
+import fixture from '../fixtures/2.json';
 
 describe('Display', () => {
   it('renders without an error', () => {
