+++ minimal_redux_poc/src/__tests__/components/ManifestMetadata.test.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { actions, store } from '../../../src/store';
-import ManifestMetadata from '../../../src/components/ManifestMetadata';
-import fixture from '../../fixtures/2.json';
+import { actions, store } from '../../store';
+import ManifestMetadata from '../../components/ManifestMetadata';
+import fixture from '../fixtures/2.json';
 
 describe('ManifestMetadata', () => {
   let wrapper;
