+++ minimal_redux_poc/src/__tests__/fixtures/2.json
